refactor: remove logging from the lower levels
logging is supposed to be a dependency in the future. spreading it all
over the place is not a good idea. the lower level packages should
instead just pass up the errors which we can log. the service layer
can decide if logging is needed before and after a database call.